requirements.txt
pandas
beautifulsoup4
yamlloader
psycopg2